refactor: migrate from `Cubit` to `StateNotifier`
