+++ features/starred_repos/lib/starred_repos.dart
@@ -10,4 +10,4 @@ export 'package:starred_repos/src/infrastructure/starred_repos.api.dart';
 export 'package:starred_repos/src/infrastructure/starred_repos.repo.dart';
 export 'package:starred_repos/src/infrastructure/starred_repos.storage.dart';
 export 'package:starred_repos/src/infrastructure/starred_repos_page_etags.storage.dart';
-export 'package:starred_repos/src/state/starred_repos.cubit.dart';
+export 'package:starred_repos/src/state/starred_repos.notifier.dart';
