+++ features/starred_repos/pubspec.yaml
@@ -8,15 +8,14 @@ environment:
   flutter: ">=2.10.1 <3.0.0"
 
 dependencies:
-  bloc: ^7.2.1
   dio: ^4.0.4
   flutter:
     sdk: flutter
   freezed_annotation: ^1.1.0
   sembast: ^3.1.2
+  state_notifier: ^0.7.2+1
 
 dev_dependencies:
-  bloc_test: ^8.5.0
   build_runner: ^2.1.7
   flutter_test:
     sdk: flutter
